fixed ft_strtrim, but ft_split - segfault

+++ ft_split.c
@@ -6,7 +6,7 @@
 /*   By: cerebus <cerebus@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/15 19:34:13 by cerebus           #+#    #+#             */
-/*   Updated: 2020/11/23 17:47:49 by cerebus          ###   ########.fr       */
+/*   Updated: 2020/11/24 16:37:08 by cerebus          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 static int	ft_count_words(char const *s, char c)
 {
-	int i;
-	int words_num;
+	int	i;
+	int	words_num;
 
 	i = 0;
 	words_num = 0;
@@ -44,13 +44,12 @@ static char	**ft_free(int arr_num, char **res)
 	return (NULL);
 }
 
-static char	**ft_fill_res(char **res, char const *s, char c,
-									unsigned int words_num)
+static char	**ft_fill_res(char **res, char const *s, char c, int words_num)
 {
-	unsigned int i;
-	unsigned int start;
-	unsigned int end;
-	unsigned int arr_num;
+	int	i;
+	int	start;
+	int	end;
+	int	arr_num;
 
 	i = 0;
 	start = 0;
@@ -64,7 +63,7 @@ static char	**ft_fill_res(char **res, char const *s, char c,
 		while (s[i] != c && s[i] != '\0')
 			i++;
 		end = i - 1;
-		res[arr_num] = ft_substr(s, start, (size_t)(end - start + 1));
+		res[arr_num] = ft_substr(s, start, end - start + 1);
 		if (!(res[arr_num]))
 			return (ft_free(arr_num, res));
 		arr_num++;
@@ -75,8 +74,8 @@ static char	**ft_fill_res(char **res, char const *s, char c,
 
 char		**ft_split(char const *s, char c)
 {
-	unsigned int	words_num;
-	char			**res;
+	int		words_num;
+	char	**res;
 
 	if (!s || !c)
 		return (NULL);
